+++ src/base/misctypes.h
@@ -5,7 +5,7 @@
 
 #include "base/ints.h"
 #include "base/general.h"
-#include "base/status_fx.h"
+#include "base/containers.h"
 
 enum {dt_pass=0, dt_lock, dt_shut, dt_boss, dt_olck, dt_osht, dt_obos, dt_wall, dt_bomb, dt_walk, dt_max};
 enum {df_walktrans=0};
@@ -163,6 +163,46 @@ struct palcycle
 	byte first,count,speed;
 };
 
+#define NUM_STATUSES 256
+struct EntityStatus
+{
+	// Damage or Healing over time
+	int32_t damage;
+	word damage_rate = 30;
+	bool damage_iframes = false;
+	bool ignore_iframes = true;
+	
+	// A sprite OR tile to overlay (or 'under'lay)
+	// EX: Static effect for electrified, flame effect for burning
+	uint8_t visual_sprite;
+	int32_t visual_tile;
+	zfix visual_x, visual_y;
+	byte visual_tilewidth = 1, visual_tileheight = 1;
+	bool visual_relative = true;
+	bool visual_under;
+	bool visual_hide_sprite; // hide the enemy/hero sprite
+	
+	// A tile modifier to the sprite's tile
+	int32_t sprite_tile_mod;
+	// If non-zero, mask out the entire sprite with this color
+	byte sprite_mask_color;
+	
+	// Which status effects are cured by this effect
+	bool cures[NUM_STATUSES];
+	
+	// Changes to defenses of the affected enemy/hero
+	bounded_map<word, int16_t> defenses {edefLAST255, -1};
+	
+	// Basic engine effects
+	bool jinx_melee, jinx_item, jinx_shield;
+	bool stun, bunny;
+	
+	bool is_empty() const;
+	
+	void clear();
+	
+	bool operator==(EntityStatus const& other) const = default;
+};
 
 enum miscsprite
 {
